权限管理

+++ src/main/java/top/mczhengyi/jvs/controller/GroupController.java
@@ -1,9 +1,11 @@
 package top.mczhengyi.jvs.controller;
 
+import org.apache.shiro.SecurityUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import top.mczhengyi.jvs.bean.Group;
 import top.mczhengyi.jvs.bean.Result;
+import top.mczhengyi.jvs.service.AuthService;
 import top.mczhengyi.jvs.service.GroupService;
 import top.mczhengyi.jvs.utils.ResultUtils;
 
@@ -14,6 +16,9 @@ import java.util.List;
 public class GroupController {
     @Autowired
     GroupService groupService;
+
+    @Autowired
+    private AuthService authService;
 
     @GetMapping("/getAll")
     public Result getAll() {
@@ -29,6 +34,8 @@ public class GroupController {
 
     @PostMapping("/insert")
     public Result insertGroup(Group group) {
+        Integer uid = (Integer) SecurityUtils.getSubject().getPrincipal();
+        group.setCreatorId(uid);
         if (groupService.insertGroup(group) == 0) {
             return ResultUtils.fail();
         } else {
@@ -38,6 +45,10 @@ public class GroupController {
 
     @PutMapping("/update/{id}")
     public Result updateGroup(@PathVariable("id") Integer gid, Group group) {
+        Integer uid = (Integer) SecurityUtils.getSubject().getPrincipal();
+        if (!groupService.getUidByGid(gid).equals(uid)) {
+            return ResultUtils.fail("权限不足");
+        }
         if (groupService.updateGroup(gid, group) == 0) {
             return ResultUtils.fail();
         } else {
@@ -47,6 +58,10 @@ public class GroupController {
 
     @DeleteMapping("/delete/{id}")
     public Result deleteGroup(@PathVariable("id") Integer gid) {
+        Integer uid = (Integer) SecurityUtils.getSubject().getPrincipal();
+        if (!groupService.getUidByGid(gid).equals(uid)) {
+            return ResultUtils.fail("权限不足");
+        }
         if (groupService.deleteGroup(gid)==0) {
             return ResultUtils.fail();
         } else {
