+++ src/main/java/top/mczhengyi/jvs/controller/VideoController.java
@@ -1,9 +1,12 @@
 package top.mczhengyi.jvs.controller;
 
+import org.apache.shiro.SecurityUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 import top.mczhengyi.jvs.bean.Result;
 import top.mczhengyi.jvs.bean.Video;
+import top.mczhengyi.jvs.service.AuthService;
 import top.mczhengyi.jvs.service.VideoService;
 import top.mczhengyi.jvs.utils.ResultUtils;
 
@@ -14,6 +17,9 @@ import java.io.IOException;
 public class VideoController {
     @Autowired
     VideoService videoService;
+
+    @Autowired
+    private AuthService authService;
 
     @GetMapping("/getAll")
     public Result getAll() {
@@ -26,21 +32,28 @@ public class VideoController {
     }
 
     @PostMapping("/insert")
-    public Result insertVideo(Video video) {
+    public Result insertVideo(@RequestParam(value = "file", required = false) MultipartFile multipartFile, Video video) {
+        Integer uid = (Integer) SecurityUtils.getSubject().getPrincipal();
+        video.setUploaderId(uid);
         if (video.getName()==null) {
             return ResultUtils.fail("请填写视频名称");
         }
         if (video.getType() == null) {
             return ResultUtils.fail("请选择视频类型");
         }
-        if (video.getUploaderId() == null) {
-            return ResultUtils.fail("未指定上传者");
+        try {
+            return ResultUtils.success(videoService.saveVideo(multipartFile, video));
+        } catch (IOException e) {
+            return ResultUtils.fail("ERROR");
         }
-        return ResultUtils.success(videoService.saveVideo(video));
     }
 
     @PutMapping("/update/{vid}")
     public Result updateVideo(@PathVariable("vid") Integer vid, Video video) {
+        Integer uid = (Integer) SecurityUtils.getSubject().getPrincipal();
+        if (!videoService.getVideoUid(vid).equals(uid)) {
+            return ResultUtils.fail("权限不足");
+        }
         if (video.getName()==null && video.getType()==null  && video.getGid()==null && video.getInfo()==null) {
             return ResultUtils.fail("请传入要修改的参数");
         }
@@ -49,6 +62,10 @@ public class VideoController {
 
     @DeleteMapping("/delete/{vid}")
     public Result deleteByVid(@PathVariable("vid") Integer vid) {
+        Integer uid = (Integer) SecurityUtils.getSubject().getPrincipal();
+        if (!videoService.getVideoUid(vid).equals(uid)) {
+            return ResultUtils.fail("权限不足");
+        }
         try {
             return ResultUtils.success(videoService.deleteByVid(vid));
         } catch (IOException e) {
